Extract helper code
