+++ src/Analyzers/CSharp/Analyzers/UseCollectionExpression/CSharpUseCollectionExpressionForStackAllocDiagnosticAnalyzer.cs
@@ -3,18 +3,14 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Immutable;
-using System.Linq;
 using System.Threading;
 using Microsoft.CodeAnalysis.CodeStyle;
-using Microsoft.CodeAnalysis.CSharp.Analyzers.UseCollectionExpression;
 using Microsoft.CodeAnalysis.CSharp.Extensions;
 using Microsoft.CodeAnalysis.CSharp.Shared.Extensions;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Options;
-using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Text;
-using Microsoft.CodeAnalysis.UseCollectionExpression;
 
 namespace Microsoft.CodeAnalysis.CSharp.UseCollectionExpression;
 
@@ -152,94 +148,11 @@ internal sealed partial class CSharpUseCollectionExpressionForStackAllocDiagnost
         StackAllocArrayCreationExpressionSyntax expression,
         CancellationToken cancellationToken)
     {
-        // has to either be `stackalloc X[]` or `stackalloc X[const]`.
-        if (expression.Type is not ArrayTypeSyntax { RankSpecifiers: [{ Sizes: [var size] } rankSpecifier] } arrayType)
-            return default;
-
-        using var _ = ArrayBuilder<CollectionExpressionMatch>.GetInstance(out var matches);
-
-        if (size is OmittedArraySizeExpressionSyntax)
-        {
-            // `stackalloc int[]` on its own is illegal.  Has to either have a size, or an initializer.
-            if (expression.Initializer is null)
-                return default;
-        }
-        else
-        {
-            // if `stackalloc X[val]`, then it `val` has to be a constant value.
-            if (semanticModel.GetConstantValue(size, cancellationToken).Value is not int sizeValue)
-                return default;
-
-            if (expression.Initializer != null)
-            {
-                // if there is an initializer, then it has to have the right number of elements.
-                if (sizeValue != expression.Initializer.Expressions.Count)
-                    return default;
-            }
-            else
-            {
-                // if there is no initializer, we have to be followed by direct statements that initialize the right
-                // number of elements.
-
-                // This needs to be local variable like `ReadOnlySpan<T> x = stackalloc ...
-                if (expression.WalkUpParentheses().Parent is not EqualsValueClauseSyntax
-                    {
-                        Parent: VariableDeclaratorSyntax
-                        {
-                            Identifier.ValueText: var variableName,
-                            Parent.Parent: LocalDeclarationStatementSyntax localDeclarationStatement
-                        },
-                    })
-                {
-                    return default;
-                }
-
-                var currentStatement = localDeclarationStatement.GetNextStatement();
-                for (var currentIndex = 0; currentIndex < sizeValue; currentIndex++)
-                {
-                    // Each following statement needs to of the form:
-                    //
-                    //   x[...] =
-                    if (currentStatement is not ExpressionStatementSyntax
-                        {
-                            Expression: AssignmentExpressionSyntax
-                            {
-                                Left: ElementAccessExpressionSyntax
-                                {
-                                    Expression: IdentifierNameSyntax { Identifier.ValueText: var elementName },
-                                    ArgumentList.Arguments: [var elementArgument],
-                                } elementAccess,
-                            }
-                        } expressionStatement)
-                    {
-                        return default;
-                    }
-
-                    // Ensure we're indexing into the variable created.
-                    if (variableName != elementName)
-                        return default;
-
-                    // The indexing value has to equal the corresponding location in the result.
-                    if (semanticModel.GetConstantValue(elementArgument.Expression, cancellationToken).Value is not int indexValue ||
-                        indexValue != currentIndex)
-                    {
-                        return default;
-                    }
-
-                    // this looks like a good statement, add to the right size of the assignment to track as that's what
-                    // we'll want to put in the final collection expression.
-                    matches.Add(new(expressionStatement, UseSpread: false));
-                    currentStatement = currentStatement.GetNextStatement();
-                }
-            }
-        }
-
-        if (!UseCollectionExpressionHelpers.CanReplaceWithCollectionExpression(
-                semanticModel, expression, skipVerificationForReplacedNode: true, cancellationToken))
-        {
-            return default;
-        }
-
-        return matches.ToImmutable();
+        return UseCollectionExpressionHelpers.TryGetMatches(
+            semanticModel,
+            expression,
+            static e => e.Type,
+            static e => e.Initializer,
+            cancellationToken);
     }
 }
